+++ dev-packages/node-integration-tests/suites/tracing/httpIntegration/instrument.mjs
@@ -1,5 +1,5 @@
-const { loggingTransport } = require('@sentry-internal/node-integration-tests');
-const Sentry = require('@sentry/node');
+import * as Sentry from '@sentry/node';
+import { loggingTransport } from '@sentry-internal/node-integration-tests';
 
 Sentry.init({
   dsn: 'https://public@dsn.ingest.sentry.io/1337',
@@ -39,20 +39,3 @@ Sentry.init({
     }),
   ],
 });
-
-// express must be required after Sentry is initialized
-const express = require('express');
-const cors = require('cors');
-const { startExpressServerAndSendPortToRunner } = require('@sentry-internal/node-integration-tests');
-
-const app = express();
-
-app.use(cors());
-
-app.get('/test', (_req, res) => {
-  res.send({ response: 'response 1' });
-});
-
-Sentry.setupExpressErrorHandler(app);
-
-startExpressServerAndSendPortToRunner(app);
